set stateToken on options so correct Router chosen
